change cluster provide name for elemental + add string to translation files for MachinePool + working on elemental cluster creation (wip - make sure save action on edit has changes on the machineinventoryselectortemplate)
