+++ pkg/elemental/types.js
@@ -2,6 +2,7 @@ export const ELEMENTAL_PRODUCT_NAME = 'elemental';
 export const ELEMENTAL_STORE = 'elemental';
 
 export const ELEMENTAL_PRODUCT_GROUP = 'elemental.cattle.io';
+export const ELEMENTAL_CLUSTER_PROVIDER = 'machineinventoryselectortemplate';
 
 export const ELEMENTAL_TYPES = {
   DASHBOARD:           'dashboard',
