+++ shell/edit/provisioning.cattle.io.cluster/MachinePool.vue
@@ -192,7 +192,7 @@ export default {
       :machine-pools="machinePools"
       @error="e=>errors = e"
     />
-    <Banner v-else color="info" label="You do not have access to see this machine pool's configuration." />
+    <Banner v-else color="info" label-key="cluster.machinePool.noAccessBanner" />
 
     <AdvancedSection :mode="mode" class="advanced">
       <portal-target :name="'advanced-' + uuid" multiple />
