+++ shell/edit/provisioning.cattle.io.cluster/rke2.vue
@@ -14,7 +14,7 @@ import {
   SCHEMA,
   DEFAULT_WORKSPACE
 } from '@shell/config/types';
-import { ELEMENTAL_SCHEMA_IDS, KIND } from '@pkg/elemental/types';
+import { ELEMENTAL_SCHEMA_IDS, KIND, ELEMENTAL_CLUSTER_PROVIDER } from '@pkg/elemental/types';
 import { _CREATE, _EDIT, _VIEW } from '@shell/config/query-params';
 
 import { findBy, removeObject, clear } from '@shell/utils/array';
@@ -263,6 +263,8 @@ export default {
   },
 
   data() {
+    console.log('******** RKE2 PROVIDER *******', this.provider);
+
     if ( !this.value.spec.rkeConfig ) {
       set(this.value.spec, 'rkeConfig', {});
     }
@@ -327,7 +329,7 @@ export default {
     },
 
     isElementalCluster() {
-      return this.provider === 'elemental' || this.value?.machineProvider === KIND.MACHINE_INV_SELECTOR_TEMPLATES.toLowerCase();
+      return this.provider === ELEMENTAL_CLUSTER_PROVIDER || this.value?.machineProvider?.toLowerCase() === KIND.MACHINE_INV_SELECTOR_TEMPLATES.toLowerCase();
     },
 
     advancedTitleAlt() {
@@ -398,7 +400,7 @@ export default {
       const showK3s = k3s.length && !existingRke2;
       const out = [];
 
-      if ( showRke2 && this.provider !== 'elemental' ) {
+      if ( showRke2 && !this.isElementalCluster ) {
         if ( showK3s ) {
           out.push({ kind: 'group', label: this.t('cluster.provider.rke2') });
         }
@@ -927,7 +929,13 @@ export default {
 
       if ( existing?.length ) {
         for ( const pool of existing ) {
-          const type = `${ CAPI.MACHINE_CONFIG_GROUP }.${ pool.machineConfigRef.kind.toLowerCase() }`;
+          let type;
+
+          if (this.isElementalCluster) {
+            type = ELEMENTAL_SCHEMA_IDS.MACHINE_INV_SELECTOR_TEMPLATES;
+          } else {
+            type = `${ CAPI.MACHINE_CONFIG_GROUP }.${ pool.machineConfigRef.kind.toLowerCase() }`;
+          }
           let config;
 
           if ( this.$store.getters['management/canList'](type) ) {
@@ -1140,6 +1148,7 @@ export default {
     },
 
     async saveOverride(btnCb) {
+      console.log('*********** OBJECT ON SAVE RKE2!!!!!! **************', this.value);
       if ( this.errors ) {
         clear(this.errors);
       }
